+++ src/Math/Combinatorics/Kostka.hs
@@ -1,17 +1,15 @@
 {-|
-Module      : Math.Algebra.Combinatorics
+Module      : Math.Combinatorics.Kostka
 Description : 
 Copyright   : (c) Stéphane Laurent, 2024
 License     : GPL-3
 Maintainer  : laurent_step@outlook.fr
 
-This module provides some functions to compute Kostka numbers with a Jack
-parameter, possibly skew, some functions to enumerate semistandard tableaux,
-possibly skew, with a given shape and a given weight, and a function to 
-enumerate the Gelfand-Tsetlin patterns defined by a skew partition.
+This module provides some functions to compute Kostka-Jack numbers, i.e.
+Kostka numbers with a Jack parameter, possibly skew.
 -}
 
-module Math.Algebra.Combinatorics
+module Math.Combinatorics.Kostka
   (
   -- * Kostka numbers
     kostkaNumbersWithGivenLambda
@@ -20,20 +18,11 @@ module Math.Algebra.Combinatorics
   , symbolicKostkaNumbers
   , skewKostkaNumbers
   , symbolicSkewKostkaNumbers
-  -- * Tableaux
-  , semiStandardTableauxWithGivenShapeAndWeight
-  , skewTableauxWithGivenShapeAndWeight
-  -- * Gelfand-Tsetlin patterns
-  , skewGelfandTsetlinPatterns
   ) where
-import qualified Data.Foldable                    as DF
 import           Data.Map.Strict                  ( 
                                                     Map
                                                   )
 import qualified Data.Map.Strict                  as DM
-import           Data.Tuple.Extra                 ( 
-                                                    second  
-                                                  )
 import           Math.Algebra.Hspray              (
                                                     RatioOfQSprays
                                                   , unitRatioOfSprays
@@ -48,18 +37,12 @@ import           Math.Algebra.Jack.Internal       (
                                                   , isSkewPartition
                                                   , skewJackInMSPbasis
                                                   , skewSymbolicJackInMSPbasis
-                                                  , _skewGelfandTsetlinPatterns
-                                                  , _skewTableauxWithGivenShapeAndWeight
-                                                  , _semiStandardTableauxWithGivenShapeAndWeight
-                                                  )
-import           Math.Combinat.Tableaux.Skew      (
-                                                    SkewTableau (..)
                                                   )
 
 -- | Kostka numbers \(K_{\lambda,\mu}(\alpha)\) with Jack parameter, or 
 -- Kostka-Jack numbers, for a given integer partition \(\lambda\) and a 
--- given Jack parameter \(\alpha\) (these are the ordinary Kostka numbers when
--- \(\alpha=1\)). This returns a map whose keys represent the 
+-- given Jack parameter \(\alpha\). These are the ordinary Kostka numbers when
+-- \(\alpha=1\). The function returns a map whose keys represent the 
 -- partitions \(\mu\) and the value attached to a partition \(\mu\) is the
 -- Kostka-Jack number \(K_{\lambda,\mu}(\alpha)\). The 
 -- partition \(\mu\) is included in the keys of this map if and only if 
@@ -83,8 +66,8 @@ kostkaNumbersWithGivenLambda lambda alpha
 -- | Kostka numbers \(K_{\lambda,\mu}(\alpha)\) with Jack parameter, or 
 -- Kostka-Jack numbers, for a given weight of the 
 -- partitions \(\lambda\) and \(\mu\) and a given Jack parameter 
--- \(\alpha\) (these are the standard Kostka numbers when
--- \(\alpha=1\)). This returns a map whose keys represent the 
+-- \(\alpha\). These are the standard Kostka numbers when
+-- \(\alpha=1\). The function returns a map whose keys represent the 
 -- partitions \(\lambda\) and the value attached to a partition \(\lambda\)
 -- represents the map \(\mu \mapsto K_{\lambda,\mu}(\alpha)\) where the 
 -- partition \(\mu\) is included in the keys of this map if and only if 
@@ -105,9 +88,9 @@ kostkaNumbers weight alpha
   | otherwise =
       _kostkaNumbers weight weight alpha 'P'
 
--- | Kostka-Jack numbers \(K_{\lambda,\mu}(\alpha)\) with symbolic Jack parameter \(\alpha\) 
--- for a given weight of the partitions \(\lambda\) and \(\mu\). This returns a map 
--- whose keys represent the 
+-- | Kostka-Jack numbers \(K_{\lambda,\mu}(\alpha)\) with symbolic Jack 
+-- parameter \(\alpha\) for a given weight of the partitions \(\lambda\) 
+-- and \(\mu\). This function returns a map whose keys represent the 
 -- partitions \(\lambda\) and the value attached to a partition \(\lambda\)
 -- represents the map \(\mu \mapsto K_{\lambda,\mu}(\alpha)\) where the 
 -- partition \(\mu\) is included in the keys of this map if and only if 
@@ -125,7 +108,7 @@ symbolicKostkaNumbers weight
 
 -- | Kostka-Jack numbers \(K_{\lambda,\mu}(\alpha)\) with symbolic Jack 
 -- parameter \(\alpha\) for a given integer partition \(\lambda\). This 
--- returns a map whose keys represent the 
+-- function returns a map whose keys represent the 
 -- partitions \(\mu\) and the value attached to a partition \(\mu\) is the
 -- Kostka-Jack number \(K_{\lambda,\mu}(\alpha)\). The 
 -- partition \(\mu\) is included in the keys of this map if and only if 
@@ -163,7 +146,7 @@ skewKostkaNumbers alpha lambda mu
 
 -- | Skew Kostka numbers \(K_{\lambda/\mu, \nu}(\alpha)\) with symbolic Jack 
 -- parameter \(\alpha\) for a given skew partition \(\lambda/\mu\). 
--- This returns a map whose keys represent the partitions \(\nu\).
+-- This function returns a map whose keys represent the partitions \(\nu\).
 symbolicSkewKostkaNumbers ::
      Partition -- ^ outer partition of the skew partition
   -> Partition -- ^ inner partition of the skew partition
@@ -174,45 +157,3 @@ symbolicSkewKostkaNumbers lambda mu
   | otherwise = 
       DM.map snd (skewSymbolicJackInMSPbasis 'P' lambda mu)
 
--- | Skew Gelfand-Tsetlin patterns defined by a skew partition and a weight vector.
-skewGelfandTsetlinPatterns :: 
-     Partition -- ^ outer partition of the skew partition
-  -> Partition -- ^ inner partition of the skew partition
-  -> [Int]     -- ^ weight
-  -> [[Partition]]
-skewGelfandTsetlinPatterns lambda mu weight 
-  | not (isSkewPartition lambda mu) =
-     error "skewGelfandTsetlinPatterns: invalid skew partition."
-  | otherwise = 
-      map (map DF.toList) (_skewGelfandTsetlinPatterns lambda mu weight)
-
--- | Skew semistandard tableaux with a given shape (a skew partition) and
--- a given weight vector. The weight is the vector whose @i@-th element is the 
--- number of occurrences of @i@ in the tableau.
-skewTableauxWithGivenShapeAndWeight :: 
-     Partition -- ^ outer partition of the skew partition
-  -> Partition -- ^ inner partition of the skew partition
-  -> [Int]     -- ^ weight
-  -> [SkewTableau Int] 
-skewTableauxWithGivenShapeAndWeight lambda mu weight 
-  | not (isSkewPartition lambda mu) =
-     error "skewTableauxWithGivenShapeAndWeight: invalid skew partition."
-  | otherwise = 
-      map (SkewTableau . (map (second DF.toList)))
-        (_skewTableauxWithGivenShapeAndWeight lambda mu weight)
-
--- | Semistandard tableaux with a given shape (an integer partition) and
--- a given weight vector. The weight is the vector whose @i@-th element is the 
--- number of occurrences of @i@ in the tableau.
-semiStandardTableauxWithGivenShapeAndWeight :: 
-     Partition   -- ^ shape, integer partition
-  -> [Int]       -- ^ weight
-  -> [[[Int]]]
-semiStandardTableauxWithGivenShapeAndWeight lambda weight 
-  | not (_isPartition lambda) =
-      error "semiStandardTableauxWithGivenShapeAndWeight: invalid partition."
-  | any (< 0) weight =
-      []
-  | otherwise = 
-      map (map DF.toList) 
-        (_semiStandardTableauxWithGivenShapeAndWeight lambda weight)